refactor: extract statistics and serialization from summary_builder
Extract pure computation functions into statistics.py (~277 lines) and
summary payload serialization into summary_serialization.py (~272 lines).
summary_builder.py shrinks from 1149 to ~704 lines and now focuses on
orchestration only.

Closes #708
