+++ apps/server/tests/use_cases/diagnostics/test_coverage_gap_audit_top10.py
@@ -29,11 +29,11 @@ from vibesensor.use_cases.diagnostics.order_analysis import (
     suppress_engine_aliases as _suppress_engine_aliases,
 )
 from vibesensor.use_cases.diagnostics.phase_segmentation import DrivingPhase
-from vibesensor.use_cases.diagnostics.summary_builder import (
-    build_phase_timeline as _build_phase_timeline,
+from vibesensor.use_cases.diagnostics.statistics import (
+    compute_accel_statistics as _compute_accel_statistics,
 )
 from vibesensor.use_cases.diagnostics.summary_builder import (
-    compute_accel_statistics as _compute_accel_statistics,
+    build_phase_timeline as _build_phase_timeline,
 )
 from vibesensor.use_cases.run import RunRecorder, RunRecorderConfig
 from vibesensor.use_cases.run.sample_builder import extract_strength_data, resolve_speed_context
